Fix lora weight fetch in isSameBatch, in viewer.html
In the 'viewer.html' file, this update corrects the comparison of 'lora' weight by fetching it from the correct object. Previously, regardless of the object in context, the weight was being fetched inappropriately solely from object 'a'. This might have resulted in inaccurate comparisons when the source object should have been 'b'.